Add TestDynamicUnary

+++ modules/Nncase.Modules.CPU/Passes/Tile/TileOptions.cs
@@ -15,7 +15,7 @@ namespace Nncase.Passes.Tile;
 /// <param name="TargetTileSize">TargetTileSize.</param>
 /// <param name="Hierarchy">the hierarchy shapes.</param>
 /// <param name="HierarchySizes">each hierarchy ram size.</param>
-public sealed record TileOptions(int[] TargetTileSize, int[] Hierarchy, int[] HierarchySizes)
+public sealed record TileOptions(int[] TargetTileSize, int[] Hierarchy, long[] HierarchySizes)
 {
-    public static TileOptions Default { get; } = new(Array.Empty<int>(), new[] { 1 }, new[] { 64 * (int)MathF.Pow(2, 30) });
+    public static TileOptions Default { get; } = new(Array.Empty<int>(), new[] { 1 }, new[] { 64 * (long)MathF.Pow(2, 30) });
 }
